Reorganize project to work with Arduino IDE
