+++ Micromouse/Encoder.cpp
